+++ R/predict.R
@@ -15,6 +15,15 @@
 #' @param min_date,max_date Optional. Date-like objects specifying the start
 #'   and end of the prediction range. See **Details** for more information on
 #'   their usage.
+#' @param day_of_week How to handle day-of-week effects when predicting
+#'   `obs_cases`. Defaults to `TRUE`, which identifies and applies the fitted
+#'   day-of-week effect, if possible. When automatic detection fails or
+#'   different levels are desirable, custom levels can be applied with a vector
+#'   of equal length to the number of desired dates. If `FALSE`, the day-of-week
+#'   effect is turned off (i.e., set to zero). When predicting parameters other
+#'   than `obs_cases` or `object` is an `RtGam` model that did not include
+#'   day-of-week effects, the day-of-week effect is turned off and this argument
+#'   is silently ignored.
 #' @param n An integer specifying the number of posterior samples to use
 #'   for predictions. Default is 100.
 #' @param mean_delay Optional. An integer specifying the mean number of days
@@ -72,11 +81,12 @@
 #' the `gratia` package. The model estimates basis function coefficients on the
 #' smooth terms \eqn{\hat \beta} and smoothing parameter(s) \eqn{\lambda}. The
 #' coefficients have the joint posterior distribution
-#' \deqn{\beta | \lambda \sim N(\hat \beta, \mathbf{V}_{\hat \beta}}
+#' \deqn{\beta | \lambda \sim N(\hat \beta, \mathbf{V}_{\hat \beta})}
 #' where \eqn{\mathbf{V}_{\hat \beta}} is the smoothing-parameter uncertainty
 #' corrected covariance matrix of the basis function coefficients. We draw
 #' samples from \eqn{\mathbf{V}_{\hat \beta}} and multiply them by the dates of
-#' interest to generate posterior draws.
+#' interest to generate posterior draws. If estimating `"Rt"` or `"r"`
+#' the day-of-week effect is excluded (i.e., set to zero).
 #'
 #' For the intrinsic growth rate, we draw one day before and one day after every
 #' day of interest. We difference these two days within the smooth to get growth
@@ -92,6 +102,7 @@
 #' For observed incident cases, we apply the estimated negative binomial
 #' observation error to the posterior expected incident cases to generate
 #' posterior predicted incident cases.
+#'
 #'
 #' @seealso [gratia::fitted_samples()], [gratia::posterior_samples()]
 #' @references Miller, David L. "Bayesian views of generalized additive
@@ -132,6 +143,7 @@ predict.RtGam <- function(
     horizon = NULL,
     min_date = NULL,
     max_date = NULL,
+    day_of_week = TRUE,
     n = 100,
     mean_delay = NULL,
     gi_pmf = NULL,
@@ -148,6 +160,7 @@ predict.RtGam <- function(
       horizon = horizon,
       min_date = min_date,
       max_date = max_date,
+      day_of_week = day_of_week,
       n = n,
       mean_delay = mean_delay,
       gi_pmf = gi_pmf,
@@ -195,6 +208,7 @@ predict_obs_cases <- function(
     horizon = NULL,
     min_date = NULL,
     max_date = NULL,
+    day_of_week = TRUE,
     n = 10,
     gi_pmf = NULL,
     seed = 12345,
@@ -206,19 +220,32 @@ predict_obs_cases <- function(
     min_date = min_date,
     max_date = max_date,
     horizon = horizon,
+    day_of_week = day_of_week,
     call = call
   )
 
-  # Use `posterior_samples()` over `fitted_samples()` to get response
-  # w/ obs uncertainty
-  fitted <- gratia::posterior_samples(
-    object[["model"]],
+  args <- list(
+    model = object[["model"]],
     data = newdata,
-    unconditional = TRUE,
     n = n,
     seed = seed,
+    unconditional = TRUE,
+    newdata.guaranteed = TRUE,
     ...
   )
+  # Turn off day of week only if requested and also
+  # was set to on in the model fitting
+  if (
+    is.factor(object[["data"]][["day_of_week"]]) && rlang::is_false(day_of_week)
+  ) {
+    args[["exclude"]] <- "s(day_of_week)"
+  }
+  fitted <- suppress_factor_warning(do.call(
+    what = gratia::posterior_samples,
+    args = args
+  ))
+  fitted[".response"] <- as.integer(fitted[[".response"]])
+
   format_predicted_dataframe(fitted, newdata)
 }
 
@@ -250,15 +277,23 @@ predict_growth_rate <- function(
   # transparent. We have 1, 2, 3, ... and map to 1, NA, 2, NA, ...
   # to show that the timesteps are (t0, t2), (t1, t3), ...
   newdata[".row"] <- interleave(seq(1, nrow(newdata) / 2, 1), NA)
-  fitted <- gratia::fitted_samples(
-    object[["model"]],
+  args <- list(
+    model = object[["model"]],
     data = newdata,
     n = n,
     seed = seed,
     unconditional = TRUE,
     scale = "linear_predictor",
+    newdata.guaranteed = TRUE,
     ...
   )
+  if (is.factor(object[["data"]][["day_of_week"]])) {
+    args[["exclude"]] <- "s(day_of_week)"
+  }
+  fitted <- suppress_factor_warning(do.call(
+    what = gratia::fitted_samples,
+    args = args
+  ))
   timestep_first_row <- which((fitted[[".row"]] - 1) %% 2 == 0)
   fitted <- data.frame(
     .row = (fitted[timestep_first_row, ".row"] + 1) / 2,
@@ -294,15 +329,23 @@ predict_rt <- function(
     gi_pmf = gi_pmf,
     call = call
   )
-  fitted <- gratia::fitted_samples(
-    object[["model"]],
+  args <- list(
+    model = object[["model"]],
     data = newdata,
     n = n,
     seed = seed,
     unconditional = TRUE,
     scale = "response",
+    newdata.guaranteed = TRUE,
     ...
   )
+  if (is.factor(object[["data"]][["day_of_week"]])) {
+    args[["exclude"]] <- "s(day_of_week)"
+  }
+  fitted <- suppress_factor_warning(do.call(
+    what = gratia::fitted_samples,
+    args = args
+  ))
   # Rt calculation
   fitted <- rt_by_group(fitted,
     group_cols = ".draw",
@@ -413,6 +456,7 @@ create_newdata_dataframe <- function(
     min_date,
     max_date,
     horizon,
+    day_of_week,
     mean_delay,
     gi_pmf,
     delta = delta,
@@ -449,9 +493,32 @@ create_newdata_dataframe <- function(
     )
   }
 
-  newdata <- gratia::data_slice(object[["model"]], timestep = timesteps)
+
+  args <- list(
+    object = object[["model"]],
+    timestep = timesteps
+  )
+
+  # Handle day of week explictly to work around gratia::data_slice
+  # needing an explicit level
+  if (is.factor(object[["data"]][["day_of_week"]])) {
+    # Use a placeholder value to satisy gratia bc it can't
+    # reliably pick a factor level without erroring
+    args["day_of_week"] <- gratia::ref_level(object[["data"]][["day_of_week"]])
+  }
+
+  newdata <- do.call(what = gratia::data_slice, args = args)
   newdata[".row"] <- seq_along(timesteps)
   newdata["reference_date"] <- desired_dates
+
+  # And re-add day of week correctly for case where it's needed
+  # All other situations will ignore day of week in drawing samples
+  if (parameter == "obs_cases") {
+    if (!rlang::is_false(day_of_week)) {
+      dow <- extract_dow_for_predict(object, day_of_week, desired_dates, call)
+      newdata["day_of_week"] <- dow
+    }
+  }
 
   newdata
 }
@@ -520,4 +587,104 @@ rt_by_group <- function(df, group_cols, value_col, vec) {
   })
 
   do.call(rbind, grouped_rt)
+}
+
+#' Manually map day-of-week effects, imputing where possible and outputting
+#' an informative error message where not.
+#' @noRd
+extract_dow_for_predict <- function(object, day_of_week, desired_dates, call) {
+  validate_day_of_week(day_of_week)
+
+  fit_data <- object[["data"]]
+  matching_dates <- desired_dates %in% fit_data[["reference_date"]]
+  all_desired_dates_in_fit <- all(matching_dates)
+  is_custom_vec <- !rlang::is_true(day_of_week) && !rlang::is_false(day_of_week)
+  if (
+    rlang::is_true(day_of_week) && all_desired_dates_in_fit
+  ) {
+    desired_dates <- fit_data[
+      fit_data[["reference_date"]] %in% desired_dates,
+      c("day_of_week", "reference_date")
+    ]
+    # dedupe in case of groups
+    deduped <- unique(desired_dates)
+    as.factor(deduped[["day_of_week"]])
+  } else if (
+    rlang::is_true(object[["day_of_week"]]) && rlang::is_true(day_of_week)
+  ) {
+    # If using default, we can impute day of week pattern
+    set_day_of_week_factor(object[["day_of_week"]], desired_dates)
+  } else if (
+    is_custom_vec && length(day_of_week) == length(desired_dates)
+  ) {
+    # If user passes an appropriate length vector and is using
+    # a custom day of week, use their input
+
+    # First check that all the levels were used in the fit
+    factor_dow <- as.factor(day_of_week)
+    known <- factor_dow %in% object[["data"]][["day_of_week"]]
+    if (!all(known)) {
+      cli::cli_abort(c(
+        "{.arg day_of_week} provided unknown level",
+        "!" = "Known levels: {.val {unique(fit_data[['day_of_week']])}}",
+        "x" = "Provided but unknown: {.val {unique(day_of_week[!known])}}"
+      ), call = call)
+    }
+    # If yes, return user-provided custom vector
+    factor_dow
+  } else if (!rlang::is_true(day_of_week)) {
+    # If vector provided but malformed error out with fix
+    n <- length(desired_dates)
+    mind <- min(desired_dates)
+    maxd <- max(desired_dates)
+    missing <- desired_dates[which(!matching_dates)]
+
+    cli::cli_abort(c(
+      "{.arg day_of_week} was provided a vector of the wrong length",
+      ">" = "Provide {.val {n}} values, for {.val {mind}} to {.val {maxd}}",
+      "x" = "Was provided {.val {length(day_of_week)}} values instead"
+    ))
+  } else {
+    # Error out with informative message about custom day of week levels
+    n <- length(desired_dates)
+    mind <- min(desired_dates)
+    maxd <- max(desired_dates)
+    missing <- desired_dates[which(!matching_dates)]
+
+    # Else tell user what we need
+    cli::cli_abort(c(
+      "{.arg day_of_week} required when using custom levels and new dates",
+      "x" = "{.val {missing}} {?wasn't/weren't} in the call to {.fn RtGam}",
+      "i" = "Pass {.fn predict} {.arg day_of_week} a vector of values to use",
+      ">" = "Provide {.val {n}} values, for {.val {mind}} to {.val {maxd}}"
+    ))
+  }
+}
+
+#' Suppress mgcv warning that factor levels are far from the data
+#' "factor levels 1 not in original fit". This warning is spurious.
+#' We're explicitly excluding the day-of-week random effect smooth but
+#' need to have some placeholder value in the dataset.
+#' @noRd
+suppress_factor_warning <- function(expr, call = rlang::caller_env()) {
+  rlang::try_fetch(
+    expr = expr,
+    warning = function(cnd) {
+      cnd_text <- as.character(cnd)
+      matched <- grepl(
+        pattern = "factor level?[s] [0-9]* not in original fit",
+        x = cnd_text
+      )
+      # Rethrow
+      if (matched) {
+        # https://stackoverflow.com/questions/49817935
+        invokeRestart("muffleWarning")
+      } else {
+        cli::cli_warn("Problem while drawing samples",
+          parent = cnd,
+          call = call
+        )
+      }
+    }
+  )
 }