+++ tests/testthat/test-dataset_creator.R
@@ -1,6 +1,7 @@
 test_that("Dataframe constructed appropriately", {
   cases <- c(1L, 2L, 3L)
   reference_date <- as.Date(c("2023-01-01", "2023-01-02", "2023-01-03"))
+  day_of_week <- FALSE
   timestep <- c(0, 0.5, 1)
 
   # Without groups
@@ -8,11 +9,18 @@ test_that("Dataframe constructed appropriately", {
     cases = cases,
     timestep = c(0, 0.5, 1),
     reference_date = reference_date,
-    group = rep(NA, 3)
+    group = rep(NA, 3),
+    day_of_week = FALSE
   )
   class(expected) <- c("RtGam_gam", class(expected))
 
-  actual <- dataset_creator(cases, reference_date, NULL, "gam")
+  actual <- dataset_creator(
+    cases = cases,
+    reference_date = reference_date,
+    group = NULL,
+    day_of_week = day_of_week,
+    backend = "gam"
+  )
   expect_equal(actual, expected)
 
   # With groups
@@ -21,11 +29,17 @@ test_that("Dataframe constructed appropriately", {
     cases = cases,
     timestep = c(0, 0.5, 1),
     reference_date = reference_date,
-    group = group
+    group = group,
+    day_of_week = FALSE
   )
   class(expected) <- c("RtGam_gam", class(expected))
 
-  actual <- dataset_creator(cases, reference_date, group, backend = "gam")
+  actual <- dataset_creator(cases,
+    reference_date,
+    group,
+    day_of_week,
+    backend = "gam"
+  )
   expect_equal(actual, expected)
 })
 
@@ -53,4 +67,37 @@ test_that("Converts double vectors to integers with a warning", {
   expect_equal(actual, integer_vec)
   expect_no_message(integerify_cases(integer_vec))
   expect_equal(integerify_cases(integer_vec), integer_vec)
+})
+
+test_that("Day of week parsed appropriately", {
+  reference_date <- as.Date(c("2023-01-01", "2023-01-02", "2023-01-3"))
+
+  # Case 1: day_of_week is TRUE so dates are auto-parsed
+  expected <- as.factor(format(reference_date, "%A"))
+  expect_equal(
+    set_day_of_week_factor(TRUE, reference_date),
+    expected
+  )
+
+  # Case 2: day_of_week is FALSE -- rep vector of FALSE
+  expected <- rep(FALSE, length(reference_date))
+  expect_equal(
+    set_day_of_week_factor(FALSE, reference_date),
+    expected
+  )
+
+  # Case 3: Manual vector
+  day_of_week_manual <- c("Weekday", "Weekend", "Holiday")
+  expected <- as.factor(day_of_week_manual)
+
+  expect_equal(
+    set_day_of_week_factor(day_of_week_manual, reference_date),
+    expected
+  )
+
+  # Case 4: Surprise unreachable
+  expect_snapshot(
+    set_day_of_week_factor(NULL, reference_date),
+    error = TRUE
+  )
 })