+++ tests/testthat/test-print.R
@@ -45,3 +45,20 @@ test_that("print method produces output for non-adaptive", {
 
   expect_snapshot(print(mock_RtGam))
 })
+
+test_that("Print method prints day of week effect", {
+  mock_RtGam <- list(
+    backend = "MockBackend",
+    m = 1,
+    k = 5,
+    model = list(family = list(family = "poisson", link = "log")),
+    data = data.frame(
+      reference_date = as.Date("2020-01-01") + 1:10,
+      group = c(rep("a", 5), rep("b", 5)),
+      day_of_week = as.factor(1:10)
+    )
+  )
+  class(mock_RtGam) <- "RtGam"
+
+  expect_snapshot(print(mock_RtGam))
+})